apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "groq>=1.0.0",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
shared = {workspace = true}

[tool.hatch.build.targets.wheel]
package = ["src/api"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "tests/",
    "README.md"
]
